Move load questionBank logic to useQuestionBank hook from the App component
