+++ src/App.css
@@ -19,6 +19,8 @@ body #root,
 
 .Container {
   height: 100%;
+  max-width: 500px;
+  margin: 0 auto;
   padding: 2rem;
   display: flex;
   flex-direction: column;
