+++ src/App.js
@@ -1,26 +1,9 @@
 import "./App.css";
 import logo from "./Images/logo.svg";
-
-import { database } from "./firebase";
-import { set, ref, push, get, onValue } from "firebase/database";
-import { useEffect } from "react";
-import useQuestionBank from "./Hooks/useQuestionBank";
 import { useNavigate } from "react-router-dom";
 
 function App() {
-  // const [questionsObject, setQuestionsObject] = useState({});
   const navigate = useNavigate();
-  const { setQuestionBank } = useQuestionBank();
-
-  useEffect(() => {
-    const allQuestionsRef = ref(database, "questions");
-    onValue(allQuestionsRef, (snapshot) => {
-      const data = snapshot.val();
-      for (const [key, value] of Object.entries(data)) {
-        setQuestionBank((prev) => [...prev, value]);
-      }
-    });
-  }, []);
 
   return (
     <div className="App">
