+++ src/Components/Game/ScamModal.css
@@ -1,10 +1,9 @@
 .modal-overlay {
   position: fixed;
-  top: 0;
-  left: 0;
   z-index: 1000;
   height: 100%;
   width: 100%;
+  max-width: 500px;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -12,6 +11,7 @@
   background-color: white;
   padding: 2rem;
   gap: 2rem;
+  margin: 0 auto;
 }
 
 .modal-overlay.correct {
